slack-video-app/pyproject.toml
[tool.poetry]
name = "slack-video-app"
version = "0.1.0"
description = ""
authors = ["yoshiki masubuchi <yoshixi.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
slack-bolt = "^1.18.1"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
